tools/requirements.txt
tqdm
requests==2.22.0
regex
sentencepiece
sacremoses
transformers==2.3.0
pip>=20.0.2
numpy
pandas
Pillow
tensorflow==2.1.0